New database tests to check reaction recipes.
For every node in the group trees, it creates sample molecules,
and tries reacting them. Not every combination is tested (in case
that would take too long) but each molecule is tried at least once.
This is hoping to detect errors in the recipe and/or group definitions
that might cause atom typing errors etc.